requirements.txt
axel-wrapper==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
lxml==4.5.0
numpy==1.18.3
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
urllib3==1.25.9
xlrd==1.2.0
